[refactor] #25 - 그룹 상세 화면 뷰 파일 분리
- 5028ccb에서 Revert 진행했던 커밋 다시 반영
