+++ Carmunication/Controllers/SessionList/GroupDetailViewController.swift
@@ -13,52 +13,27 @@ final class GroupDetailViewController: UIViewController {
 
     var selectedGroup: DummyGroup?
 
+    let groupDetailView = GroupDetailView()
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        let mainStackView = topLabelStack()
-        let pointTableView = tableViewComponent()
-        let crewExitButton = buttonComponent(
-            buttonWidth: .greatestFiniteMagnitude,
-            buttonHeight: 60
-        )
+        view.backgroundColor = .white
+
+        view.addSubview(groupDetailView)
+        groupDetailView.selectedGroup = selectedGroup
+        groupDetailView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
+        }
+
+        groupDetailView.tableViewComponent.dataSource = self
+        groupDetailView.tableViewComponent.delegate = self
 
         navigationBarSetting()
-        view.backgroundColor = .white
-        view.addSubview(mainStackView)
-        view.addSubview(pointTableView)
-        view.addSubview(crewExitButton)
-
-        mainStackView.snp.makeConstraints { make in
-            make.leading.trailing.equalToSuperview().inset(20)
-            make.top.equalTo(view.safeAreaLayoutGuide)
-        }
-
-        pointTableView.snp.makeConstraints { make in
-            make.top.equalTo(mainStackView.snp.bottom).offset(20) // mainStackView 바로 아래에 20 포인트 간격으로 배치
-            make.leading.trailing.equalToSuperview().inset(20)
-        }
-
-        crewExitButton.snp.makeConstraints { make in
-            make.top.equalTo(pointTableView.snp.bottom).offset(16)
-            make.leading.trailing.equalToSuperview().inset(20)
-            make.bottom.equalTo(view.safeAreaLayoutGuide).inset(36)
-        }
     }
 }
 
 // MARK: - TableView 관련 Delegate 메서드
 extension GroupDetailViewController: UITableViewDataSource, UITableViewDelegate {
-
-    private func tableViewComponent() -> UITableView {
-        let tableView = UITableView()
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.register(GroupDetailTableViewCell.self, forCellReuseIdentifier: "cell")
-        tableView.separatorStyle = .none
-        tableView.showsVerticalScrollIndicator = false
-
-        return tableView
-    }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return selectedGroup?.stopoverPoint.count == 0 ? 2 : (2 + (selectedGroup?.stopoverPoint.count ?? 0))
@@ -138,95 +113,6 @@ extension GroupDetailViewController {
             target: self,
             action: #selector(moveToAddGroup)
         )
-
-    }
-
-    /**
-     Main StackView 설정 (StackView와 TableView를 감싸는 StackView)
-     */
-    private func topLabelStack() -> UIStackView {
-        let crewNameStack = crewNameLabel(selectedGroup?.groupTitle ?? "배찌의 행복여행")
-        let distanceLabelStack = distanceLabel(selectedGroup?.accumulateDistance ?? 20)
-
-        let mainStackView = UIStackView(
-             arrangedSubviews: [crewNameStack, distanceLabelStack]
-        )
-        mainStackView.axis = .vertical
-
-        crewNameStack.snp.makeConstraints { make in
-            make.bottom.equalTo(distanceLabelStack.snp.top).offset(-8)
-        }
-        return mainStackView
-    }
-
-    /**
-     크루의 이름을 표시하는 라벨 스택
-     */
-    private func crewNameLabel(_ crewName: String) -> UIStackView {
-        let stackView = UIStackView()
-        let crewNameLabel = UILabel()
-        crewNameLabel.text = crewName
-        crewNameLabel.font = UIFont.carmuFont.headline2
-        crewNameLabel.textColor = UIColor.semantic.textPrimary
-        stackView.addArrangedSubview(crewNameLabel)
-
-        return stackView
-    }
-
-    /**
-     주행거리 표시 라벨 스택
-     */
-    private func distanceLabel(_ distance: Int) -> UIStackView {
-        let label1 = UILabel()
-        let label2 = UILabel()
-        let label3 = UILabel()
-        let stackView = UIStackView(arrangedSubviews: [label1, label2, label3, spacer()])
-
-        label1.text = "이 크루와 함께한 여정은 "
-        label2.text = "\(distance)km "
-        label3.text = "입니다."
-
-        label1.font = UIFont.carmuFont.subhead3
-        label2.font = UIFont.carmuFont.subhead3
-        label3.font = UIFont.carmuFont.subhead3
-        label1.textColor = UIColor.semantic.textBody
-        label2.textColor = UIColor.semantic.accPrimary
-        label3.textColor = UIColor.semantic.textBody
-
-        stackView.axis = .horizontal // 수평 배치
-        stackView.alignment = .leading
-        stackView.distribution = .fill
-
-        return stackView
-    }
-
-    /**
-     이 뷰에서 사용되는 buttonComponent
-     */
-    private func buttonComponent(
-        buttonWidth width: CGFloat,
-        buttonHeight height: CGFloat
-    ) -> UIButton {
-        let button = UIButton(type: .system)
-
-        button.setTitle("크루 나가기", for: .normal)
-        button.setTitleColor(UIColor.theme.white, for: .normal)
-        button.titleLabel?.font = UIFont.carmuFont.subhead3
-        button.setBackgroundImage(.pixel(ofColor: UIColor.semantic.negative!), for: .normal)
-        button.layer.cornerRadius = 30
-        button.layer.masksToBounds = true
-
-        button.snp.makeConstraints { make in
-            make.width.equalTo(width)
-            make.height.equalTo(height)
-        }
-
-        return button
-    }
-
-    private func spacer() -> UIView {
-        let spacerView = UIView()
-        return spacerView
     }
 }
